soc: pinctrl-lpi: Fix core hw vote issue in LPI GPIO access
Fix core hw vote issue by adding retry logic in
retrieving clock handle. Since the LPI driver is an active
listener to SSR notifications, it gets an early notification
as soon as DSP is up. So in order to avoid access
of LPASS registers of LPI GPIOs before AVS services are up,
check for AVS up, vote for clock and access GPIOs.

Change-Id: I2aca83154d24a3a97af4b5b0b449257f77ab99c1
Signed-off-by: Aditya Bavanari <abavanar@codeaurora.org>