net: qualcomm: rmnet: Segmentation cleanup
There is a lot of repeated code in the various segmentation functions.
Introduce and use a struct to hold various pieces of metadata we'll need
later on during the segmentation process like the lengths and positions of
headers. Passing this struct allows us to only compute various values once
instead of in every function, and cleans up the parameter lists.

Change-Id: I19d36701573c818ddbd8d3f8d1c1ace717712bec
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>