arm64: hibernate: Introduce new entry point to kernel
Introduce a new entry point to hibernated kernel image.
This is generally needed when bootloader restores the
hibernated image from disc to ddr and passes control
to it by turning off the mmu, also initialize this new
entry point with cpu_resume which turns on the mmu and
then proceeds with restore routines.

Change-Id: I854880e244910614153996a1f9ea923124369ae2
Signed-off-by: Vivek Kumar <vivekuma@codeaurora.org>