net: qualcomm: rmnet: check for over-pulling
When pulling data from the packet descriptor or when trimming the
descriptor to a specific length, passing invalid or overly large sizes
can result in the descriptor being recycled as it no longer contains any
packet data. We must check for this situation occurring, and gracefully
handle the packet drop.

Change-Id: Idb1354105904b0fff265a7ef32fee8905f7825ae
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>