net: qrtr: Add support for a USB device transport
Add a transport that uses the USB device ipc_bridge APIs.
The USB device IPC APIs are used to transfer data from a
device to the host over USB. The modem will generally be
configured as a device and this transport is intended to
be used on the modem controller co-processor.

Change-Id: I8871fef55b266bb4f749e7963bf12e9fe1ae7281
Signed-off-by: Gustavo Solaira <gustavos@codeaurora.org>