msm: mhi_dev: Avoid re-alloc of netdev interface
If netdev interface_init gets called by mhi,
when the channels are not actually closed.
trying to do a reinit results in error from mhi
(like re-opening a already opened channel).
that leads to kernel crashes.
Add a check to avoid re-allocating netdev interface if it
has been already allocated.

Change-Id: I3835312a3e1fb09e979cc1dcf0b23815d77b7b57
Signed-off-by: Siva Kumar Akkireddi <sivaa@codeaurora.org>