msm: npu: Fix missing interrupt issue when sending packet to firmware
It is possible that qhdr_rx_req is changed by firmware while
driver is writing the command packet to the IPC queue. Then driver
may not send an interrupt to fw at the end but fw does require
an interrupt. This change is to fix this potential racing condition
by checking if an interrupt is requested by firmware after
qhdr_write_idx is updated.

Change-Id: I9c5af690543607e8b3c71eed9120360063f93fb4
Signed-off-by: Da Hoon Pyun <dpyun@codeaurora.org>