ANDROID: cuttlefish_defconfig: Disable CMOS RTC driver
Android expects rtc0 to support saving the system clock, however the
CMOS rtc emulated by crosvm currently only supports 24hr alarms, so
it is not useful. As this useless rtc driver occupied rtc0 it broke
various tests and features. The older crosvm used by cuttlefish did
not simulate an RTC so it was not affected because the driver probe
failed.

Bug: 158000323
Change-Id: I54590ceb0f09e6b1ff1f7ed094dab66ed7e58672
Signed-off-by: Alistair Delva <adelva@google.com>