usb: host: xhci-plat: Fix IMOD configuration from sysfs
Currently config_imod_store function checks for the return value
of kstrtouint and bails out if it is not 1. But the success case
return value of kstrtouint is 0 instead, with -ERANGE and -EINVAL
being error case values. This leads to the IMOD setting failure.
Fix this by fixing the return value check of kstrtouint.

Change-Id: I723f4efbe9c39fc91d5f57ef657daaab61fba212
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>