usb: host: xhci: Increase number of event ring segments to 4
Currently the event ring has just a single segment consisting of
256 TRBs. When features like ECM which have very fast completions
due to short packet sizes are exercised, then the CPU is not able
to drain the event ring as fast as the USB HW fills it up with
new completion events. This leads to USB HW finding the event
ring to be full which can result in watchdog error.
Fix this by increasing the number of ER segments to 4.

Change-Id: Icc0ed65e5042179729315a8fd96f87b8b99e87f3
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>