usb: dwc3-msm: Allow SSPHYs to enter P3 irrespective of device connect
Currently, the dual port controller's SSPHYs autosuspend is
disabled when USB is in resumed state. They are allowed to go
into P3 only if a SuperSpeed device is connected to one of the
ports and suspend is triggered. This leads to SSPHYs not in P3
state when suspend happens without any SuperSpeed device
connected. Consequently, any SS device connect is not detected.
Fix this by allowing SSPHYs to go into P3 by allowing autosuspend
irrespective of device connect.

Change-Id: Ia614a1778f8e2837b368e8ac959e330434e5bac3
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>