arm64: entry: Add vectors that have the bhb mitigation sequences
commit ba2689234be92024e5635d30fe744f4853ad97db upstream.

Some CPUs affected by Spectre-BHB need a sequence of branches, or a
firmware call to be run before any indirect branch. This needs to go
in the vectors. No CPU needs both.

While this can be patched in, it would run on all CPUs as there is a
single set of vectors. If only one part of a big/little combination is
affected, the unaffected CPUs have to run the mitigation too.

Create extra vectors that include the sequence. Subsequent patches will
allow affected CPUs to select this set of vectors. Later patches will
modify the loop count to match what the CPU requires.

Change-Id: I9d37133eeb0038e2297c42a9f4db677407b27c33
Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Git-commit: 41a7b3dad498b6ba3db8c4ab802c64ab3e89167d
Git-repo: https://android.googlesource.com/kernel/common/
[quic_c_spathi@quicinc.com: reintroduced function
 arm64_update_smccc_conduit() to fix compilation issue]
Signed-off-by: Srinivasarao Pathipati <quic_c_spathi@quicinc.com>